+++ src/designsystem/Avatar/Avatar.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Avatar } from "./Avatar";
+import { Avatar } from ".";
 
 const meta = {
   title: "Avatar",
