+++ src/designsystem/AvatarButton/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Avatar, AvatarProps } from "../Avatar/Avatar";
+import { Avatar, AvatarProps } from "../Avatar";
 
 type AvatarButtonProps = AvatarProps & React.ButtonHTMLAttributes<HTMLButtonElement>;
