+++ src/designsystem/AvatarLink/AvatarLink.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { fn } from "@storybook/test";
-import { AvatarLink } from "./AvatarLink";
+import { AvatarLink } from ".";
 
 const meta = {
   title: "AvatarLink",
